+++ build-release.sh
@@ -48,7 +48,6 @@ fi
 echo "Running tests ..."
 
 php -d zend.enable_gc=0 -d date.timezone=UTC -d mbstring.func_overload=7 -d mbstring.internal_encoding=UTF-8 vendor/bin/phpunit --coverage-text --configuration tests/travis/sqlite.phpunit.xml
-phpunit
 
 if [ $? -ne 0 ]
 then
@@ -63,19 +62,6 @@ echo "Building HTML docs; please check output for warnings ..."
 echo ""
 
 make html
-
-echo ""
-
-if [ $? -ne 0 ]
-then
-	echo "Build FAILED!"
-	exit 1
-fi
-
-echo "Building EPUB docs; please check output for warnings ..."
-echo ""
-
-make epub
 
 echo ""