+++ user_guide_src/source/changelog.rst
@@ -34,6 +34,7 @@ Bug fixes for 3.1.12
 -  Fixed a bug (#5903) - :doc:`common function <general/common_functions>` :php:func:`set_status_header()` didn't recognize 'HTTP/2.0' as a valid ``$_SERVER['SERVER_PROTOCOL']``.
 -  Fixed a bug (#6013) - :doc:`Session <libraries/sessions>` flashdata didn't work on PHP 8.
 -  Fixed a bug (#6006) - ``is_callable()`` change in PHP 8 broke :doc:`Migrations <libraries/migration>`, a part of :doc:`XML-RPC <libraries/xmlrpc>` and an edge case in 404 detection logic.
+-  Fixed a bug (#5729) - :doc:`Query Builder <database/query_builder>` possibly not detecting ``NOT BETWEEN`` expression.
 
 Version 3.1.11
 ==============