+++ user_guide_src/source/changelog.rst
@@ -152,7 +152,11 @@ Bug fixes for 3.2.0
 Version 3.1.12
 ==============
 
-Release Date: Not Released
+Release Date: Mar 3, 2022
+
+- **Security**
+
+   -  Fixed a possible session fixation vulnerability where ``session.use_strict_mode`` wasn't enforced on PHP 7+.
 
 -  General Changes